+++ zerver/tests/test_transfer.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, patch
 from django.conf import settings
 from moto import mock_s3
 
-from zerver.lib.actions import check_add_realm_emoji
+from zerver.actions.realm_emoji import check_add_realm_emoji
 from zerver.lib.avatar_hash import user_avatar_path
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import (
